requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.1.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.2.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
cudatoolkit=11.3.1=h9edb442_10
cycler=0.11.0=pypi_0
datasets=2.10.1=pypi_0
diff-match-patch=20230430=pypi_0
dill=0.3.5.1=pypi_0
docker-pycreds=0.4.0=pypi_0
evaluate=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.10.4=h0708190_1
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
huggingface-hub=0.13.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
javalang=0.13.0=pypi_0
jbig=2.1=h7f98852_2003
joblib=1.2.0=pypi_0
jpeg=9e=h166bdaf_1
kiwisolver=1.4.4=pypi_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
lerc=2.2.1=h9c3ff4c_0
libdeflate=1.7=h7f98852_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=hf544144_1
libuv=1.43.0=h7f98852_0
libwebp-base=1.2.2=h7f98852_1
lxml=4.9.2=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
matplotlib=3.5.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h402132d_0
mkl_fft=1.3.1=py37h3e078e5_1
mkl_random=1.2.2=py37h219a48f_0
multidict=6.0.4=pypi_0
multiprocess=0.70.13=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.6=he412f7d_0
nltk=3.6.7=pypi_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
olefile=0.46=pyh9f0ad1d_1
openai=0.27.6=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pypi_0
pandas=1.1.5=pypi_0
pathtools=0.1.2=pypi_0
pillow=8.3.2=py37h0f21c89_0
pip=22.3.1=py37h06a4308_0
portalocker=2.7.0=pypi_0
prettytable=3.7.0=pypi_0
promise=2.3=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
pyarrow=11.0.0=pypi_0
pyparsing=3.0.9=pypi_0
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
python_abi=3.7=2_cp37m
pytorch=1.10.1=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.2=pypi_0
responses=0.18.0=pypi_0
rouge=1.0.1=pypi_0
rouge-score=0.0.4=pypi_0
sacrebleu=2.3.1=pypi_0
sacremoses=0.0.53=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
sentencepiece=0.1.97=pypi_0
sentry-sdk=1.16.0=pypi_0
seqeval=1.2.2=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py37h06a4308_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pypi_0
sqlite=3.40.1=h5082296_0
tabulate=0.9.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
torchaudio=0.10.1=py37_cu113
torchvision=0.11.2=py37_cu113
tqdm=4.65.0=pypi_0
transformers=4.17.0=pypi_0
typing_extensions=4.4.0=pyha770c72_0
urllib3=1.26.15=pypi_0
wandb=0.12.20=pypi_0
wcwidth=0.2.6=pypi_0
wheel=0.38.4=py37h06a4308_0
xxhash=3.2.0=pypi_0
xz=5.2.10=h5eee18b_1
yarl=1.8.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.0=ha95c52a_0
